scripts/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
requests==2.26.0
python-dotenv==0.19.0
Werkzeug==3.0.1
asyncpg==0.26.0
psycopg2==2.9.7